refactor(cosmos-staking): move positions previews to PreviewActivity
Preview composables don't belong in the main source set — the @Preview
functions + their mock data compiled into the production APK. Per the
project convention (and #4806), screen previews live in the debug-only
PreviewActivity.

- Drop the three @Preview functions (loaded / loading / skeleton) and the
  preview-state helpers from CosmosStakingPositionsScreen; make
  CosmosStakingPositionsContent and StakingPositionSkeleton internal so
  debug can call them.
- Re-add the previews in PreviewActivity, wired into the screen dispatcher
  as cosmos_staking_positions_loaded / _loading / _position_skeleton.
