+++ app/src/debug/java/com/vultisig/wallet/debug/PreviewActivity.kt
@@ -39,6 +39,7 @@ import androidx.compose.ui.unit.sp
 import com.vultisig.wallet.R
 import com.vultisig.wallet.data.blockchain.cosmos.qbtc.claim.QbtcClaimBlockedReason
 import com.vultisig.wallet.data.blockchain.cosmos.qbtc.claim.QbtcClaimError
+import com.vultisig.wallet.data.blockchain.cosmos.staking.CosmosStakePositionRow
 import com.vultisig.wallet.data.models.Address
 import com.vultisig.wallet.data.models.Chain
 import com.vultisig.wallet.data.models.Coins
@@ -70,6 +71,7 @@ import com.vultisig.wallet.ui.models.ChainTokensUiModel
 import com.vultisig.wallet.ui.models.TransactionDetailsUiModel
 import com.vultisig.wallet.ui.models.TransactionScanStatus
 import com.vultisig.wallet.ui.models.VerifyTransactionUiModel
+import com.vultisig.wallet.ui.models.cosmosstaking.CosmosStakingPositionsUiState
 import com.vultisig.wallet.ui.models.cosmosstaking.CosmosStakingVerifyUiState
 import com.vultisig.wallet.ui.models.cosmosstaking.CosmosStakingVerifyValidatorRow
 import com.vultisig.wallet.ui.models.deposit.DepositFormUiModel
@@ -92,7 +94,9 @@ import com.vultisig.wallet.ui.models.swap.ValuedToken
 import com.vultisig.wallet.ui.models.swap.VerifySwapUiModel
 import com.vultisig.wallet.ui.models.toNetworkUiModel
 import com.vultisig.wallet.ui.screens.TransactionDoneView
+import com.vultisig.wallet.ui.screens.cosmosstaking.CosmosStakingPositionsContent
 import com.vultisig.wallet.ui.screens.cosmosstaking.CosmosStakingVerifyContent
+import com.vultisig.wallet.ui.screens.cosmosstaking.StakingPositionSkeleton
 import com.vultisig.wallet.ui.screens.deposit.BondFormContent
 import com.vultisig.wallet.ui.screens.keygen.FastVaultVerificationScreen
 import com.vultisig.wallet.ui.screens.keygen.ImportSeedphraseContent
@@ -128,6 +132,7 @@ import dagger.hilt.EntryPoint
 import dagger.hilt.InstallIn
 import dagger.hilt.android.EntryPointAccessors
 import dagger.hilt.components.SingletonComponent
+import java.math.BigDecimal
 
 @EntryPoint
 @InstallIn(SingletonComponent::class)
@@ -216,6 +221,11 @@ class PreviewActivity : ComponentActivity() {
                     "staking_verify_after" -> CosmosStakingVerifyCtaPreview(newButtons = true)
                     "staking_verify_qbtc" ->
                         CosmosStakingVerifyCtaPreview(newButtons = true, qbtc = true)
+                    "cosmos_staking_positions_loaded" ->
+                        CosmosStakingPositionsContentLoadedPreview()
+                    "cosmos_staking_positions_loading" ->
+                        CosmosStakingPositionsContentLoadingPreview()
+                    "cosmos_staking_position_skeleton" -> StakingPositionSkeletonPreview()
                     "sign_message_before" -> VerifySignMessageCtaPreview(newButtons = false)
                     "sign_message_after" -> VerifySignMessageCtaPreview(newButtons = true)
                     else -> SwapConfirmPreview()
@@ -1910,5 +1920,116 @@ private fun ColumnScope.VerifySignMessageCta(newButtons: Boolean) {
             variant = VsButtonVariant.Secondary,
             modifier = Modifier.fillMaxWidth(),
         )
+    }
+}
+
+private fun cosmosStakingPreviewPosition(
+    address: String,
+    moniker: String,
+    staked: String,
+    fiat: String,
+    reward: String,
+    status: CosmosStakePositionRow.ValidatorStatus,
+    apy: BigDecimal? = BigDecimal("0.121"),
+) =
+    CosmosStakePositionRow(
+        validatorAddress = address,
+        validatorMoniker = moniker,
+        validatorIdentity = null,
+        stakedAmount = BigDecimal(staked),
+        stakedFiatDisplay = fiat,
+        pendingReward = BigDecimal(reward),
+        apyPercent = apy,
+        validatorAvatarUrl = null,
+        validatorStatus = status,
+        pendingUnbondingUnlockDate = null,
+        pendingUnbondingEntryCount = 0,
+    )
+
+private fun cosmosStakingLoadedState() =
+    CosmosStakingPositionsUiState(
+        ticker = "LUNC",
+        coinLogo = "",
+        positions =
+            listOf(
+                cosmosStakingPreviewPosition(
+                    address = "terravaloper1allnodes78wk0n3d3kjm0lue0ramln8m8r2x7yg",
+                    moniker = "Allnodes",
+                    staked = "1000",
+                    fiat = "$58.00",
+                    reward = "0.512345",
+                    status = CosmosStakePositionRow.ValidatorStatus.Active,
+                ),
+                cosmosStakingPreviewPosition(
+                    address = "terravaloper1churned9wk0n3d3kjm0lue0ramln8m8r2zzzz",
+                    moniker = "Churned Val",
+                    staked = "234.5",
+                    fiat = "$13.60",
+                    reward = "0.0",
+                    status = CosmosStakePositionRow.ValidatorStatus.ChurnedOut,
+                    apy = null,
+                ),
+            ),
+        hasClaimableRewards = true,
+        totalStaked = BigDecimal("1234.5"),
+        totalStakedFiat = "$71.60",
+        totalAmountPrice = "$71.60",
+        selectedPositions = listOf("LUNC"),
+    )
+
+@Composable
+private fun CosmosStakingPositionsContentLoadedPreview() {
+    CosmosStakingPositionsContent(
+        state = cosmosStakingLoadedState(),
+        chainId = "TerraClassic",
+        isRefreshing = false,
+        onRefresh = {},
+        onManagePositions = {},
+        onClaim = {},
+        onDelegateToNewValidator = {},
+        onUnstake = {},
+        onMove = {},
+        onStakeMore = {},
+        onPositionSelectionChange = { _, _ -> },
+        onPositionSelectionDone = {},
+        onDismissDialog = {},
+    )
+}
+
+@Composable
+private fun CosmosStakingPositionsContentLoadingPreview() {
+    CosmosStakingPositionsContent(
+        state =
+            CosmosStakingPositionsUiState(
+                ticker = "LUNC",
+                coinLogo = "",
+                positions = emptyList(),
+                isLoading = true,
+                selectedPositions = listOf("LUNC"),
+            ),
+        chainId = "TerraClassic",
+        isRefreshing = false,
+        onRefresh = {},
+        onManagePositions = {},
+        onClaim = {},
+        onDelegateToNewValidator = {},
+        onUnstake = {},
+        onMove = {},
+        onStakeMore = {},
+        onPositionSelectionChange = { _, _ -> },
+        onPositionSelectionDone = {},
+        onDismissDialog = {},
+    )
+}
+
+@Composable
+private fun StakingPositionSkeletonPreview() {
+    Column(
+        modifier =
+            Modifier.background(Theme.v2.colors.backgrounds.primary).padding(16.dp).fillMaxWidth(),
+        verticalArrangement = Arrangement.spacedBy(12.dp),
+    ) {
+        StakingPositionSkeleton()
+        StakingPositionSkeleton()
     }
 }
